treewide: enable cpu scheduler
This mega-patch shows how to enable the CPU scheduler. It received only
scant testing (but appears to work).  It needs a lot more testing,
splitting to easily reviewable patches.

The patch focuses on background loads like compaction, flush, and repair
as in most latency sensitive loads the foreground loads are well below
capacity. Even there, there are probably some expensive continuations that
are not tagged correctly.